refactor: :truck: extract order-example
