+++ src/pages/actions.tsx
@@ -1,6 +1,6 @@
 import { useAppBridge } from "@saleor/app-sdk/app-bridge";
-import { Button, Text, Box } from "@saleor/macaw-ui/next";
-import { useOrdersQuery } from "../../generated/graphql";
+import { Box, Button, Text } from "@saleor/macaw-ui/next";
+import { OrderExample } from "../order-example";
 
 /**
  * This is example of using AppBridge, when App is mounted in Dashboard
@@ -11,23 +11,12 @@ import { useOrdersQuery } from "../../generated/graphql";
  */
 const ActionsPage = () => {
   const { appBridge, appBridgeState } = useAppBridge();
-  const [{ data, fetching }] = useOrdersQuery();
 
   const navigateToOrders = () => {
     appBridge?.dispatch({
       type: "redirect",
       payload: {
         to: "/orders",
-        actionId: "message-from-app",
-      },
-    });
-  };
-
-  const navigateToOrder = (id: string) => {
-    appBridge?.dispatch({
-      type: "redirect",
-      payload: {
-        to: `/orders/${id}`,
         actionId: "message-from-app",
       },
     });
@@ -70,35 +59,7 @@ const ActionsPage = () => {
           </Button>
         </Box>
       </Box>
-      <Box display="flex" flexDirection={"column"} gap={2}>
-        <Text as={"h2"} variant={"heading"}>
-          Fetching data
-        </Text>
-
-        <>
-          {fetching ? (
-            <Text>Fetching orders...</Text>
-          ) : (
-            <>
-              <Text color="textNeutralSubdued">
-                💡 You can modify the query in the <code>graphql/queries/Orders.graphql</code> file.
-                Remember to run <code>pnpm codegen</code> afterwards to regenerate the types.
-              </Text>
-              <ul style={{ listStyle: "inside", margin: 0 }}>
-                {data?.orders?.edges?.map((order) => (
-                  <li key={order.node.id}>
-                    <a
-                      onClick={() => navigateToOrder(order.node.id)}
-                      href={`https://pikle.eu.saleor.cloud/dashboard/orders/${order.node.id}`}
-                    >{`Order ${order.node.number}`}</a>{" "}
-                    {`is ${order.node.status}`}
-                  </li>
-                ))}
-              </ul>
-            </>
-          )}
-        </>
-      </Box>
+      <OrderExample />
       <Box display="flex" flexDirection={"column"} gap={2}>
         <Text as={"h2"} variant={"heading"}>
           Webhooks
@@ -107,7 +68,6 @@ const ActionsPage = () => {
           The App Template contains an example <code>ORDER_CREATED</code> webhook under the path{" "}
           <code>src/pages/api/order-created</code>.
         </Text>
-
         <Text as="p">
           Create any{" "}
           <Text
